[FEAT] 레지스터 페이지 수정 #72
